refactor(frontend): split model equipment rendering
