+++ engine/examples/editor-playtest/src/model_rendering/equipment.rs
@@ -1,0 +1,324 @@
+use super::*;
+
+#[derive(Copy, Clone)]
+struct AttachmentPose {
+    origin: WorldVertex,
+    rotation: Mat3I16,
+}
+
+#[allow(clippy::too_many_arguments)]
+pub(super) fn draw_player_equipment(
+    current_room: RoomIndex,
+    character: RuntimeCharacter,
+    models: &[Option<RuntimeModelAsset>; MAX_RUNTIME_MODELS],
+    model_faces: &[TexturedModelRenderFace],
+    model_parts: &[ModelPart],
+    model_vertices: &[ModelVertex],
+    clips: &[Option<Animation<'static>>; MAX_RUNTIME_MODEL_CLIPS],
+    x: i32,
+    y: i32,
+    z: i32,
+    yaw: Angle,
+    anim_action: CharacterAnimationAction,
+    clip_local: ModelClipIndex,
+    anim_start_tick: SimTick,
+    elapsed_tick: SimTick,
+    video_hz: VideoHz,
+    camera: &WorldCamera,
+    options: WorldSurfaceOptions,
+    lighting: &RuntimeRoomLighting,
+    triangles: &mut impl PrimitiveSink<TriTextured>,
+    world: &mut WorldRenderPass<'_, '_, OT_DEPTH>,
+) -> EquipmentDrawStats {
+    let mut out = EquipmentDrawStats::default();
+    let Some(character_model) = models.get(character.model.to_usize()).copied().flatten() else {
+        return out;
+    };
+    let Some(character_anim) = character_model.clip(clips, clip_local) else {
+        return out;
+    };
+    let local_tick = elapsed_tick.saturating_sub(anim_start_tick);
+    let character_phase = animation_phase_at_tick_q12(
+        character_anim,
+        local_tick,
+        video_hz,
+        character.action_loops(anim_action),
+    );
+    let character_anchor = model_clip_anchor(character_model, clip_local);
+    let reference_anchor = model_clip_anchor(character_model, character.clip_for(PlayerAnim::Idle));
+    let character_pose_translation = model_pose_anchor_translation(
+        character_anim,
+        character_phase,
+        character_anchor,
+        reference_anchor,
+        character.action_in_place_override(anim_action),
+    );
+    let character_frame = (character_phase >> 12) as u16;
+    let character_model_rotation = yaw_rotation_matrix(yaw.add_signed_q12(character.visual_yaw));
+    let character_origin = visual_model_origin(
+        x,
+        y,
+        z,
+        character_model.world_height,
+        character.visual_offset,
+        character.visual_scale_q8,
+        &character_model_rotation,
+    );
+    let character_local_to_world =
+        visual_model_local_to_world(character_model, character.visual_scale_q8);
+
+    let mut drawn = 0usize;
+    for equipment in EQUIPMENT {
+        if equipment.room != current_room
+            || equipment.flags & equipment_flags::PLAYER == 0
+            || drawn >= MAX_EQUIPMENT_DRAWS
+        {
+            continue;
+        }
+        let Some(weapon) = WEAPONS.get(equipment.weapon.to_usize()) else {
+            continue;
+        };
+        let Some(socket) = find_model_socket(character_model, equipment.character_socket)
+            .or_else(|| find_model_socket(character_model, weapon.default_character_socket))
+        else {
+            continue;
+        };
+        let Some(socket_pose) = attachment_socket_pose(
+            character_model,
+            character_anim,
+            character_phase,
+            character_origin,
+            character_model_rotation,
+            character_local_to_world,
+            character_pose_translation,
+            socket,
+        ) else {
+            continue;
+        };
+        let weapon_rotation = socket_pose
+            .rotation
+            .mul(&euler_q12_rotation_inverse(weapon.grip_rotation_q12));
+
+        match weapon.model {
+            Some(model_index) => {
+                let Some(weapon_model) = models.get(model_index.to_usize()).copied().flatten()
+                else {
+                    continue;
+                };
+                let grip = scaled_offset(weapon_model.local_to_world, weapon.grip_translation);
+                let grip_world = rotate_offset_q12(&weapon_rotation, grip);
+                let origin = WorldVertex::new(
+                    socket_pose.origin.x.saturating_sub(grip_world[0]),
+                    socket_pose.origin.y.saturating_sub(grip_world[1]),
+                    socket_pose.origin.z.saturating_sub(grip_world[2]),
+                );
+                if let Some(anim) = weapon_model.clip(clips, weapon_model.default_clip) {
+                    let phase = anim.phase_at_tick_q12(elapsed_tick.as_u32(), video_hz.as_u16());
+                    let material = lighting.shade_model_material(origin, weapon_model.material);
+                    let model_options = options
+                        .with_depth_policy(DepthPolicy::Average)
+                        .with_cull_mode(CullMode::Back)
+                        .with_material_layer(material)
+                        .with_textured_triangle_splitting(true)
+                        .with_textured_triangle_max_edge(MODEL_TEXTURE_SPLIT_MAX_EDGE);
+                    let faces = runtime_model_faces(weapon_model, model_faces);
+                    let stats = submit_runtime_model_predecoded(
+                        world,
+                        triangles,
+                        weapon_model,
+                        anim,
+                        phase,
+                        *camera,
+                        origin,
+                        weapon_rotation,
+                        weapon_model.local_to_world,
+                        ModelPoseTranslation::ZERO,
+                        material,
+                        model_options,
+                        faces,
+                        model_parts,
+                        model_vertices,
+                    );
+                    accumulate_model_stats(&mut out.stats, stats);
+                    if stats.primitive_overflow || stats.command_overflow {
+                        out.draws = drawn as u16;
+                        return out;
+                    }
+                    drawn += 1;
+                    out.draws = drawn as u16;
+                }
+            }
+            None => {}
+        };
+
+        let (active, hits) = evaluate_weapon_hitboxes(
+            current_room,
+            weapon.hitbox_first.to_usize(),
+            weapon.hitbox_count,
+            character_frame,
+            socket_pose.origin,
+            socket_pose.rotation,
+        );
+        out.active_hitboxes = out.active_hitboxes.saturating_add(active);
+        out.target_hits = out.target_hits.saturating_add(hits);
+    }
+    out
+}
+
+fn find_model_socket(
+    model: RuntimeModelAsset,
+    name: &str,
+) -> Option<&'static LevelModelSocketRecord> {
+    let first = model.socket_first.to_usize();
+    let count = model.socket_count as usize;
+    let sockets = MODEL_SOCKETS.get(first..first.saturating_add(count))?;
+    sockets.iter().find(|socket| socket.name == name)
+}
+
+fn attachment_socket_pose(
+    _model: RuntimeModelAsset,
+    animation: Animation<'static>,
+    phase_q12: u32,
+    origin: WorldVertex,
+    instance_rotation: Mat3I16,
+    local_to_world: LocalToWorldScale,
+    pose_translation: ModelPoseTranslation,
+    socket: &LevelModelSocketRecord,
+) -> Option<AttachmentPose> {
+    let pose = apply_model_pose_translation(
+        animation.pose_looped_q12(phase_q12, socket.joint)?,
+        pose_translation,
+    );
+    let joint = compute_joint_world_transform(pose, instance_rotation, local_to_world, origin);
+    Some(compose_socket_pose(
+        joint,
+        socket.translation,
+        socket.rotation_q12,
+    ))
+}
+
+fn compose_socket_pose(
+    joint: JointWorldTransform,
+    translation: [i32; 3],
+    rotation_q12: [i16; 3],
+) -> AttachmentPose {
+    let offset = rotate_offset_q12(&joint.rotation, translation);
+    let local_rotation = euler_q12_rotation(rotation_q12);
+    AttachmentPose {
+        origin: WorldVertex::new(
+            joint.translation.x.saturating_add(offset[0]),
+            joint.translation.y.saturating_add(offset[1]),
+            joint.translation.z.saturating_add(offset[2]),
+        ),
+        rotation: joint.rotation.mul(&local_rotation),
+    }
+}
+
+fn evaluate_weapon_hitboxes(
+    current_room: RoomIndex,
+    first: usize,
+    count: u16,
+    frame: u16,
+    origin: WorldVertex,
+    rotation: Mat3I16,
+) -> (u16, u16) {
+    let mut active = 0u16;
+    let mut hits = 0u16;
+    let Some(hitboxes) = WEAPON_HITBOXES.get(first..first.saturating_add(count as usize)) else {
+        return (0, 0);
+    };
+    for hitbox in hitboxes {
+        if frame < hitbox.active_start_frame || frame > hitbox.active_end_frame {
+            continue;
+        }
+        active = active.saturating_add(1);
+        for entity in ENTITIES {
+            if entity.room != current_room {
+                continue;
+            }
+            if weapon_hit_shape_hits_point(hitbox.shape, origin, rotation, entity.x, entity.z) {
+                hits = hits.saturating_add(1);
+            }
+        }
+    }
+    (active, hits)
+}
+
+fn weapon_hit_shape_hits_point(
+    shape: WeaponHitShapeRecord,
+    origin: WorldVertex,
+    rotation: Mat3I16,
+    px: i32,
+    pz: i32,
+) -> bool {
+    match shape {
+        WeaponHitShapeRecord::Box {
+            center,
+            half_extents,
+        } => {
+            let c = transform_local_point(origin, rotation, center);
+            let radius = half_extents[0].max(half_extents[2]) as i32;
+            distance_xz_sq(RoomPoint::new(px, 0, pz), RoomPoint::new(c.x, 0, c.z))
+                <= square_i32_saturating(radius)
+        }
+        WeaponHitShapeRecord::Capsule { start, end, radius } => {
+            let a = transform_local_point(origin, rotation, start);
+            let b = transform_local_point(origin, rotation, end);
+            point_segment_xz_distance_sq(px, pz, a.x, a.z, b.x, b.z)
+                <= square_i32_saturating(radius as i32)
+        }
+    }
+}
+
+fn transform_local_point(origin: WorldVertex, rotation: Mat3I16, point: [i32; 3]) -> WorldVertex {
+    let offset = rotate_offset_q12(&rotation, point);
+    WorldVertex::new(
+        origin.x.saturating_add(offset[0]),
+        origin.y.saturating_add(offset[1]),
+        origin.z.saturating_add(offset[2]),
+    )
+}
+
+fn scaled_offset(scale: LocalToWorldScale, offset: [i32; 3]) -> [i32; 3] {
+    [
+        scale.apply(offset[0]),
+        scale.apply(offset[1]),
+        scale.apply(offset[2]),
+    ]
+}
+
+fn euler_q12_rotation(rotation_q12: [i16; 3]) -> Mat3I16 {
+    let rx = Mat3I16::rotate_x(Angle::from_q12(rotation_q12[0] as u16).rotate_y_arg());
+    let ry = Mat3I16::rotate_y(Angle::from_q12(rotation_q12[1] as u16).rotate_y_arg());
+    let rz = Mat3I16::rotate_z(Angle::from_q12(rotation_q12[2] as u16).rotate_y_arg());
+    rz.mul(&ry).mul(&rx)
+}
+
+fn euler_q12_rotation_inverse(rotation_q12: [i16; 3]) -> Mat3I16 {
+    let inv_x = (-(rotation_q12[0] as i32)) as u16;
+    let inv_y = (-(rotation_q12[1] as i32)) as u16;
+    let inv_z = (-(rotation_q12[2] as i32)) as u16;
+    let rx = Mat3I16::rotate_x(Angle::from_q12(inv_x).rotate_y_arg());
+    let ry = Mat3I16::rotate_y(Angle::from_q12(inv_y).rotate_y_arg());
+    let rz = Mat3I16::rotate_z(Angle::from_q12(inv_z).rotate_y_arg());
+    rx.mul(&ry).mul(&rz)
+}
+
+fn point_segment_xz_distance_sq(px: i32, pz: i32, ax: i32, az: i32, bx: i32, bz: i32) -> i32 {
+    let abx = bx.saturating_sub(ax);
+    let abz = bz.saturating_sub(az);
+    let apx = px.saturating_sub(ax);
+    let apz = pz.saturating_sub(az);
+    let denom = square_i32_saturating(abx).saturating_add(square_i32_saturating(abz));
+    if denom <= 0 {
+        return square_i32_saturating(apx).saturating_add(square_i32_saturating(apz));
+    }
+    let dot = apx
+        .saturating_mul(abx)
+        .saturating_add(apz.saturating_mul(abz));
+    let t_q8 = ratio_q8_i32(dot.clamp(0, denom), denom);
+    let cx = ax.saturating_add((abx.saturating_mul(t_q8)) >> 8);
+    let cz = az.saturating_add((abz.saturating_mul(t_q8)) >> 8);
+    square_i32_saturating(px.saturating_sub(cx))
+        .saturating_add(square_i32_saturating(pz.saturating_sub(cz)))
+}
